added more to the Swift Testing unit tests, wroking on cleaning up model bugs

+++ RYSplitViewSearchExample/ViewModes/ViewModel.swift
@@ -11,16 +11,35 @@ import SwiftData
 
 @Observable @MainActor
 class ViewModel {
-    // saved items in swiftData
-    var savedItems = [Item]()
+    // Saved items in SwiftData
+    private(set) var savedItems = [Item]()
 
-    // results from searching by name with searchText
+    // An 'Item' from `savedItems` array that is selected and displayed in the details view, if `searchedItem` is not set.
+    var selectedItem: Item?
+    
+    // The result of the user's`searchText` query
+    var searchedItem: Item?
+    
+    // The current search text entered by the user.
+    var searchText: String = ""
+    
+    // if the search bar is selected/activated.
+    var isSearchPresented: Bool = false
+    
+    // Indicates if `searchedItem` search result is being displayed in a modal Sheet (iOS only)
+    var isDetailSheetPresented: Bool = false
+    
+    // Used for Search results and Auto-complete search results of `Items` for the current `searchText` value
+    // Case-insensitive search if `searchText` text is contianed within the `name` property of `searchableItems`
     var searchedItemsByName: [Item] {
         searchableItems.filter { $0.name.lowercased().contains(searchText.lowercased()) }
     }
 
-    //the list of searchables..
-    private var searchableItems: [Item] {
+    // The internal list of searchable Items
+    private var searchableItems = [Item]()
+
+    // initial values to use for `searchableItems`
+    private var initialSearchableItems: [Item] {
         get {
             let names = ["Apple",
                          "Banana",
@@ -43,22 +62,17 @@ class ViewModel {
                          "Papaya",
                          "Plum",
                          "Apricot"]
-            return names.sorted().map { Item(name: $0) }
+            return names.sorted().map { Item($0) }
         }
     }
-    
-    var selectedItem: Item?
-    var searchedItem: Item?
-
-    var searchText: String = ""
-    var isSearchPresented: Bool = false
-    var isDetailSheetPresented: Bool = false
     
     private var modelContext: ModelContext
 
     init(modelContext: ModelContext) {
         self.modelContext = modelContext
-        fetchSavedData()
+        fetchSavedItems()
+        searchableItems = initialSearchableItems
+        
 #if os(iOS)
         if let first = savedItems.first {
             selectedItem = first
@@ -66,10 +80,10 @@ class ViewModel {
 #endif
     }
     
-    private func fetchSavedData() {
+    private func fetchSavedItems() {
         do {
             let descriptor = FetchDescriptor<Item>(sortBy: [SortDescriptor(\.name)])
-            savedItems = try modelContext.fetch(descriptor)
+            savedItems = try modelContext.fetch(descriptor).sorted{ $0.createdAt < $1.createdAt }
         } catch {
             print("Fetch failed")
         }
@@ -77,14 +91,30 @@ class ViewModel {
     
     func add(item: Item) {
         modelContext.insert(item)
-        fetchSavedData()
         try? modelContext.save()
+        fetchSavedItems()
     }
     
-    func delete(item: Item) {
-        modelContext.delete(item)
-        fetchSavedData()
-        try? modelContext.save()
+    func delete(_ item: Item) {
+        if let index = savedItems.firstIndex(of: item) {
+            modelContext.delete(item)
+            try? modelContext.save()
+            fetchSavedItems()
+            
+#if os(macOS)
+
+            //if there still is an item at this index, select it
+            if savedItems.indices.contains(index) {
+                selectedItem = savedItems[index]
+            // if not, and there is at least one item, select the first
+            } else if let first = savedItems.first {
+                selectedItem = first
+            // nothing is selectable
+            } else {
+                selectedItem = nil
+            }
+#endif
+        }
     }
             
     func onSubmitOfSearch() {
